requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
distlib==0.4.0
dj-database-url==2.1.0
Django==4.2.16
django-cloudinary-storage==0.3.0
filelock==3.19.1
fonttools==4.61.1
freetype-py==2.5.1
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
packaging==26.0
pdfkit==1.0.0
pillow==12.1.0
platformdirs==4.4.0
psycopg2-binary==2.9.9
pycairo==1.29.0
pycparser==3.0
pydyf==0.12.1
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.6.2
PyPDF2==3.0.1
pyphen==0.17.2
python-bidi==0.6.7
python-dotenv==1.0.0
PyYAML==6.0.3
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.34.0
weasyprint==68.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.17
zopfli==0.4.0
